requirements.txt
PyMySQL~=1.1.0
pandas~=2.2.0
SQLAlchemy~=2.0.25
mysql-connector-python~=8.3.0